External data fetched for pipeline is now stored
External data fetched for pipeline is now stored in a new directory that is timestamped to the second. The purpose of this is to aid in troubleshooting. When changes to the results occur it is difficult to tell if this is due to a change in the code or a change in the external resources. Keeping a log of the external fetched data makes it possible to tell what is the cause.